Twitter.tcl: Default to polling home timeline only.
In a recent commit the default was switched from home to mentions only.
This was confusing as polling the home timeline was the original intent
of the script and how many users use it. This commit sets the default to
the most popular (I believe) use case.

I also improve a few comments in oauth.tcl, twitlib.tcl, and
twitter.tcl.