Add a config option to map followed accounts to specific channels
Previously every tweet would go to every +twitter channel. This is still
the default. With this new config you can limit certain accounts to
output to only certain channels. Any account not listed in the new
mapping behaves as it did before, going to all channels. Any account
listed in the new mapping goes only to the listed channels.