Allow !following/!followers to default to the bot's account
Previously you would have to know the bot's screen name and give it as a
parameter to these triggers. We can just default to nothing and we will
get our own results.

This also fixes a bug where the count parameter was incorrectly named.
This lead to us always seeing a maximum of 20 screen names.