Use extended tweet mode
This allows us to see more than 140 characters of tweets when that can
happen. This includes stuff like mentions and attachments. See
https://dev.twitter.com/overview/api/upcoming-changes-to-tweets for more
info. Prior to this we'd receive only the truncated text of tweets in
some cases. This came up as desirable in #23.

Also, there were a couple cases (!twit_updates and !twit_search) where
we did not run the "fix status" logic, which cleans up tweet text and
tries to get the full text of a retweeted text. This is now fixed and
should be done consistently.