Avoid exceeding ratelimit when polling every minute
Previously we'd be able to regularly exceed the accepted rate. This
happens even if polling only the home timeline. It is sufficient to
adjust our rate to every 61 seconds to avoid this.

This changes us to allow update times other than 1, 5, or 10 as well.

An alternative solution may be to look at the ratelimit headers we get
back and avoid making a request if we see there are none left. However
we still have to deal with making new requests in a timely manner if
we're right against the limit.