Don't output tweets that occurred prior to us following a user
Previously we could end up with a bunch of tweets showing up from a
newly followed user all at once. This is because the decision to show a
tweet was based on the last tweet ID we saw.

We now record the user's last tweet ID upon follow, and skip it and any
older than it from the user. Newer tweets will show up.